fix pkexec
can't detach the process. all of polkit refuses to auth a process which
has pid1 as parent. a startdetached process would be one of those
and pkexec is broken. to work around this we'll simply hold on to the
child and be their parent.

other options would be to detach sh and have the upgrader run as a fork
of that. OR exec() into the upgrader directly. latter is unclear if it is
safe to do in Qt, also if the upgrader doesn't actually finish this would
mean that the notifier isn't running until the user relogs, so it's
generally wise to keep the notifier running