implement a devel switch
the notifier is now registered on dbus org.kde.DistroReleaseNotifier
and exposes a UseDevel property which is basically a runtime config
for using the devel variant in update-manager.

changing the property causes a new poll for a new version. ideally this
would result in a notification popup. the actual upgrader is also invoked
with the devel switch based on the runtime configuration of the dbus
object.

this enables the user to call

```
qdbus org.kde.DistroReleaseNotifier / org.kde.DistroReleaseNotifier.UseDevel true
```

to trigger a poll for devel versions. the devel version poll continues to
be enabled until the app is restarted (i.e. mostly when a relog happens)

ideally we'd have another helper app in bin/ which attempts poke the devel
property. this would serve as a shorthand for the aforementioned command
(although maybe we should avoid qdbus since it's buggy as hell thanks to
qtchooser).

additionally the dbus interface has a Poll method which specifically
serves as a means to interactive force a poll outside the timer triggers.
this is mostly only useful for testing and has no real real life use IMO