add a KSNI to the notification and refactor into wrapper class
notification business is now isolated into a Notifier class.
the Notifier class wraps a KSNI + KNotification notification system.

KSNI lends us some persistence that the knotification persistance doesn't
offer because it still disappears after some seconds and then only
sits in the notification box where it's next to useless. to give us extra
visibility the KSNI is brought up at the same time and will not go away.

kbroulik says he has plans to fix the knotification persistence, but that's
not until 5.15. so this is the best way forward for now.

T9474