hide the ksni when the upgrader is running
when the upgrader is detected as running we'll delete the ksni thus hiding
it (there is no actual hide feature :().

currently we will not restore the ksni if the upgrade was aborted. this
could be done but is a bit tricky. AFAIK the upgrader creates no marker
file on successful upgrade, so we'd probably have to trigger a new check
instead of simply showing ourselves again. otherwise we'd show the
ksni again if the user already upgraded but chose not to reboot?

T9552