inhibit screensaver based on upgrader presence
this prevents the screen from locking while the upgrader is running. during
the upgrade the lockscreen may not be working correctly anymore because of
runtime qml dependencies getting lost etc. as such we shouldn't allow
locking during upgrades.

this adds a new string, i18n isn't particularly important for this though.
it is shown in the power applet popout I think