do not leak processes from previous sessions
session restore would keep restoring the previous processes along with
the autostarted process because the notifieri isn't strictly speaking
a kdbusaddons unique app. as a result we'd keep starting more and more
notifiers for each log in.

we'll now simply exit if the notifier is getting started on session restore

BUG: 399010