deal with do-release-upgrader exiting before it has a UI
for reasons beyond me there's fatal checks being done before
spinning up the UI which means the upgrade can be refused with the only
viable output method being stdout.
to deal with this we now have a upgraderprocess class that wraps around
the upgrader. should the process finish !0 before it appeared on dbus we'll
consider it failed and throw up a fairly crappy error dialog showing
the merged stdout/stderr channels.
once it appears on dbus we consider it self-sufficient and stop our
monitoring.

this sucks a fair amount but is the most reasonable way of dealing with
this. we could to a degree replicate the checks on our end to preempt
what do-release-upgrader is doing, but who's to say we'll have exactly
the same behavior. plus, how would we make sure we always new checks
that might get added.
so... all in all while the UI/UX is fairly garbage here this is ultimately
because the ubuntu release upgrader itself is being daft.
in fact, I've tried on ubuntu 19.10 and it brings up the 'yo, upgrade?'
dialog but then it falls over because of likely the very same problem that
do-release-upgrader just do random UI-less checks after that.