port to kded
the standalone binary was fairly awkward because it required its own
private heap of a bunch of underlying libraries (e.g. xcb) that caused
it to have a multi-MiB footprint.
this excess consumption goes away by moving to kded since we now only
have the private heap we actually require reducing the footprint to a
couple hundred KiB

this sees a change in the dbus api since we cannot register the object
on / any longer. instead we now register on an RDN-derived name.

moving forward we could opt to replace the dbus api with the builtin
capabilities of kdedmodule (doesn't really have an advantage though
other than getting rid of a fairly tiny class)